requirements.txt
tensorboardx>=1.4
torch>=0.4
docopt>=0.6
tensorboard>=1.8
googledrivedownloader>=0.3
torchvision>=0.2.1
Pillow>=5.3.0
matplotlib
scipy
scikit-image
sklearn
Pillow
pandas
git+git://github.com/FamousDirector/pycwt.git